Add/Delete Folder
Added the ability to add and delete folders, however, I need to
not only fix and redraw the screen to actually display the folder
changes, but I need some serious refactoring of the current
class. It's currently too large and needs to be split into atleast
a 'OpenFolder' which deals with the extensions and a 'MainView' class,
which deals with the folders themselves.