Major Refactoring
Split up the code into seperate classes which should make modification
easier to do. Also added some color to the buttons to help the user
understand the actions a little better.

Still might add an 'Are you sure' screen when the user wants to delete
a folder/extension so any accidentaly deletions can be further
prevented.